firmware: Refactor Platform RPC as Reporter method
As described in the attached bug, the Platform RPC causes test flakes.
This CL deletes the Platform RPC, which was used to report the DUT's
board and model, and rewrites it as two separate methods on the
firmware Reporter: r.Board(), and r.Model().

BUG=b:166150709
TEST=firmware.Config on a coral/nasher

Cq-Depend: chromium:2372633
Change-Id: I72d8593f4a67077c36ceaa5464e81664e56fecac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2373207
Tested-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Greg Edelston <gredelston@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>