camera/JPEG: Stop the UI prior to running test.
This CL changes camera.DecodeAccelJPEG and camera.EncodeAccelJPEG to
stop the UI prior to running the JPEG hardware acceleration unit tests.
This is not strictly necessary for the tests to work. However, if the UI
is playing a video, this could interfere with the test if there is a
driver bug that causes crashes/panics. Therefore, for stability, it's
best to stop the UI since it's unrelated to the test.

BUG=b:161878038
TEST=camera.DecodeAccelJPEG on a zork and EncodeAccelJPEG on a nami

Change-Id: I1ff98c9afc48a01148aeee0d74dcaadfbdf9670b