tast-tests: add test for DeviceMinimumVersion policy
This CL adds a test for DeviceMinimumVersion policy for the case
of critical updates i.e. zero warning time. The test verifies that
the in case of critical updates the user is logged out of the
session and an update required screen is shown on the login page.

BUG=chromium:1121550
TEST=tast run <ip> policy.DeviceMinimumVersion

Change-Id: Ieaad0bbc7a16bfa88b6144387753164808ecc422
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2377950
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Tested-by: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>