Add desktop UI components tests
This is yet another test which will run as a step of CQ to
prevent significant performance regressions. This test case
focuses on the performance of several desktop UI components,
more specifically on launcher and the quick settings.

See go/cros-ui-perftests-cq for further details of this test.

BUG=chromium:1107698
TEST=tast run ui.DesktopControl on kevin

Change-Id: Iabeabc6be6b32fe8d3b3e0b45a705be548eef798
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2486406
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>
Reviewed-by: Toni Baržić <tbarzic@chromium.org>