Fix flake in ui.ChromeCrashReporterMetrics.miss_crashpad
As part of the test, we loop and kill all the crashpad_handler
processes. However, there is a race where if a previous iteration sent a
kill signal to a crashpad_handler, but the process is only actually
reaped at the moment the next loop is executing, we try to send a second
kill signal and then fail out because the kill signal can't be
delivered.

Don't fail the loop if the kill signal can't be delivered. Instead, just
try again until we get a loop where there were no crashpad_handlers.

BUG=chromium:1103042
Fixed: chromium:1103042
TEST=tast test ui.ChromeCrashReporterMetrics.miss_crashpad on cyan

Change-Id: I129ee814a77265da299ab7ac663525c47469fe77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2377650
Tested-by: Ian Barkley-Yeung <iby@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>
Commit-Queue: Ian Barkley-Yeung <iby@chromium.org>