video.DecodeAccel: Use model instead of platform to filter for CQ.
This CL changes the video.DecodeAccel.h264_cq test to filter by model
instead of platform id. The model names are adjusted to match the
models used in Chromium CQ.

This CL also updates a comment to better clarify the choice of
devices and removes some notes that were not very relevant.

BUG=chromium:949459
TEST=video.DecodeAccel.h264_cq on kohaku

Change-Id: Idbe5d40c68e7741bfe1ab6d1bce3398e0463c862