tast-tests: appcompat limit test case time
This limits the time to run an individual test case. This helps to stop
a single failing test case from using all of the time. It can be
overridden by setting a specific timeout of the test case. Also shortens
the long ui timeout to a more reasonable value.

BUG=None
TEST=ran some of the app compat tests

Change-Id: I071016338dee38ecfa940e790b54df0981640537
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2378726
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Meenakshi Thiyagarajan <mthiyagarajan@chromium.org>
Reviewed-by: Meenakshi Thiyagarajan <mthiyagarajan@google.com>