Add RT07 Tast-test for WM ARC++ R
Implement RT07: resizable/clamshell: immerse via API from maximized.
In RC07 a resizable activity is started at tablet mode and window
bounds are stored. Then the immersive button is clicked. This should
take the activity to fullscreen mode. Then the normal button is
clicked, this should take the activity to maximize mode and the
window bounds should be equal to the original window bounds in root.

BUG=b:140203139
TEST=tast run $DUT arc.WMNonresizableTablet

Change-Id: I7380d066b5c4eec61d586457a7624386bd0bd08b