tast: Add UIInputGain to check input gain by UI control API
Add new tast "UIInputGain" to check the end-to-end controllability of
audio input node gain. Gain levels are controlled by UI API to adjust
the slider in the system tray.

The criteria is to use ALSA loopback path to record audio data by a
lower and a higher gain respectively, and check whether the ratio of RMS
amplitudes of these two audio data is as expected.

BUG=1083465
BUG=b:158857828
TEST=tast run ${IP} audio.UIInputGain

Cq-Depend: chromium:2405606
Change-Id: I219f849b3f9c4e495412ad745ff982facc25123d