Add RC03 Tast-test for WM ARC++ R
Implement RC03: resizable/clamshell: maximize non-portrait app.
In RC03 a landscape and an unspecified activities are tested. For
each activity a click to Maximize button is simulated and a check to
make sure the window is maximized is performed. Then Restore button
is clicked and window bounds are compared to the original values.
For more info please visit go/arc-wm-r.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableClamshel

Change-Id: Ibd2691e78d2e5aa03baaeb81a225ec7272c7e3e2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2388973
Tested-by: Armen Keshishian <armenk@google.com>
Reviewed-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Armen Keshishian <armenk@google.com>