arc: Update Clipboard test to retry pasting.
As per b/157615371#comment23, the "copy" message sent between android
and chrome are asynchronous and therefore cannot be expected to show up
~immediately in all cases. To handle this complexity, I added a handful
of quick retries to the majority of the test cases in this file. (And to
reduce duplication, I did a good amount of refactoring of the existing
tests.)

I'm looking for confirmation from test/feature owners that this is the
right thing to do. See comments in the bug about timing investigation (e.g.
b/157615371#comment28); in short, we rarely need to retry for longer
than the blink of an eye.

The remaining tests (CopyImageFromChromeToAndroid and
ACopyHTMLFromChromeToAndroidWithObserver) didn't get refactored because
they were structured a little differently. They only flake a little bit
according to stainless, at a lower rate than the other tests, but
we might want to loop back later and either duplicate the retries to the
other test cases, or integrate them into the table test I'm introducing.

BUG=b:157615371
TEST=`tast run $IP arc.Clipboard` many times on eve and eve-kernelnext

Change-Id: Iaf5f6b21d50d1b628f3847358cf914df24ae63e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2382458
Tested-by: Kevin Chowski <chowski@google.com>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Yuichiro Hanada <yhanada@chromium.org>
Commit-Queue: Kevin Chowski <chowski@google.com>