ZipPerf: fix folder navigation failure
Follow-up to crrev/c/2366519, after which the test has just ~1%
failures. These remaining failures seem all similar: After renaming a
folder and waiting that the input box disappears, sending an 'Enter'
keyboard event to navigate to this folder seems to still be source of
flakiness.

'Enter' keypress event is replaced with a double-click.
When double clicking on a folder, after changing location, the focus is
already on the files box. filesBox.FocusAndWait has been removed (test
fails otherwise).

BUG=chromium:407010
TEST=tast run DUT_IP filemanager.ZipPerf

Change-Id: I10841712964a62d37f39bb237860cf6d4450c760