Disable crash.ECCrash on some broken models.
Some devices have broken EC firmware and cannot report crashes. It's
infeasible to fix this, so disable the test on those boards. See
discussion in https://crbug.com/1123191 for details.

Note: This change does not disable the test on all boards listed in said
bug, only the ones that it's run (and failed) on as shown in tastboard:
https://tastboard.corp.google.com/test/crash.ECCrash

BUG=chromium:1123191,chromium:1039380
TEST=tast run $ASUKA_DUT crash.ECCrash

Change-Id: I60b95ec189df8f83e700b635769f54e40d720eaf