Crostini: wait for tablet transition in restart tests
There is a race between switching to clamshell mode and right clicking
on the shelf icon which results in the "shutdown linux" option not
appearing.

I verified this locally on hatch, and adding a wait for UI animations to
complete makes this work on hatch locally.

BUG=chromium:1123797
TEST=Ran restart test on hatch in both tablet and clamshell mode.

Change-Id: I7356d5edaeed9d085e8274d82438b865a18b0c1a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2386706
Tested-by: Matthew Warton <mwarton@google.com>
Reviewed-by: Keigo Oka <oka@chromium.org>
Reviewed-by: David Munro <davidmunro@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>