wifi: PTK: Add a stricter case.
In b/167149633, we find a problem in rekey on specific platforms and
might lose all pings in one rekey, but the test might still pass with
ping loss rate just below the failure threshold. However, having a
stricter check might increase the false alarm rate and make the test
too flaky to be informative.

Let's add another test case for debugging the problem for now and can
update the main test when we find a good set of parameter to detect
the problem.

BUG=b:167149633
TEST=tast run $DUT wifi.PTK*

Change-Id: Ib56a5cdd7acd762bed2cbe820d57d124823c39d7