tast-tests: add test for DeviceMinimumVersion policy notifications
This CL adds a test for DeviceMinimumVersion policy for the case
of end of life or auto update expiration. The test verifies that
the in case of eol, the user is is shown the in-session notification
and a message on the Chrome management page.

BUG=chromium:1121550
TEST=tast run <ip> policy.DeviceMinimumVersionNotifications

Change-Id: Id9cad1547a7c4773d3b967816d61527ded5aaa7a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2388063
Tested-by: Saurabh Nijhara <snijhara@google.com>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>
Reviewed-by: Oleh Lamzin <lamzin@google.com>