tast-tests: save the screenshot file during the test
The screenshot used to be saved during the test and it's helpful to
check the screenshot if the test fails. Due to change in
crrev.com/c/1691882, the screenshot is removed when the test finished.
To have the option to keep the screenshot files after the test, an
additional parameter is added in the screenshot.GrabScreenshot function.

Currently, arc.ResizeActivity and arc.QuarterSizedWindowZooming will
save the screenshot in this patch. It's recommended for other tests too.

BUG=b:167419204
TEST=tast -verbose run <DUT> arc.BlackFlash
TEST=tast -verbose run <DUT> arc.ResizeActivity
TEST=tast -verbose run <DUT> arc.QuarterSizedWindowZooming

Change-Id: I813dc50057861fb5ec5984d8edb06588ed76f432