tast-tests: Add hwdep.[No]ForceDischarge to PowerCameraGcaPreviewPerf
On some platforms, force discharging a battery with ectool is not
supported. If so, the battery metric should not be collected.
This CL parameterizes PowerCameraGcaPreviewPerf to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297
TEST=tast run <DUT w/o battery> arc.PowerCameraGcaPreviewPerf* did not
fail

Change-Id: Iaa15c411f01cea31e0fadadc36468d41f130fb85
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2387914
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Matthias Springer <springerm@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>