tast-tests: Add hwdep.[No]ForceDischarge to CameraPerfExtraMetrics
On some platforms, force discharging a battery with ectool is not
supported. If so, the battery metric should not be collected.
This CL parameterizes CameraPerfExtraMetrics to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297
TEST=tast run <DUT w/o battery> arc.CameraPerfExtraMetrics* did not
fail

Change-Id: I5ffe713bb15879547a9378297015dd61cd8a285e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2389801
Reviewed-by: Matthias Springer <springerm@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>