tast-tests: reduce automation click latency.
The back and forth between Go and JS causes the latency between getting
a node's location and clicking the node around ~10-60ms. This change reduces
the latency to ~5-20ms. This doesn't fix the race condition, but it
greatly reduces the window in which it can occur. This is likely as
close as we can get to atomic mouse clicks on automation nodes while
keeping the clicks E2E.

BUG=b:161438238
TEST=Made sure tests with clicking still passed.

Change-Id: Ib2bb96fe280000f83f2f6d7d8447d69033e9ded3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2390844
Tested-by: Brendan Hansknecht <bhansknecht@chromium.org>
Commit-Queue: Brendan Hansknecht <bhansknecht@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>