tast-tests: Add hwdep.[No]ForceDischarge to PowerCameraPreviewPerf
On some platforms, force discharging a battery with ectool is not
supported. If so, the battery metric should not be collected.
This CL parameterizes PowerCameraPreviewPerf to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297
TEST=tast run <DUT w/o battery> arc.PowerCameraPreviewPerf* did not
fail

Change-Id: Ieb5d6835bf3fe511daf50c55dccb7a506574280e