WIP: add test for fwupd
Now the test run success. With a dummy cab file(a metadata file with
all empty binaries) and writes the flashrom command to the output
folder. Next step could be to verify the flashrom command. So far
didn't see the requirement of using a real binary but open to discussion.

BUG=b:166206618
TEST=tast run 127.0.0.1:2222 firmware.FwupdInstallPs175

Change-Id: If3a544c686256accbcd6314f2f16696bcfa06a4e