tast-tests: arc.RemovableMedia: Support content URIs starting from UUIDs
We are changing the format of content URIs used by ArcVolumeProvider to
content://org.chromium.arc.volumeprovider/<UUID>/*.

This CL adds the support for such URIs to arc.RemovableMedia, while the
support for old content URIs is kept as is.

Also, it makes the test informational. This is because it will enter a
not-well-tested state when new URIs start to be used. We can switch it
back to critical and stop the support for old URIs after a while.

BUG=b:164934426
TEST=$ ~/trunk/src/platform/tast/fast_build.sh -T
TEST=Run the following command for DUTs using old/new content URIs:
TEST=$ tast run <DUT> arc.RemovableMedia

Change-Id: Ia9f53f180aa631a94c4795087b8a223c225d5378