tast-tests: arc.RemovableMedia: Support content URIs starting from UUIDs
We are changing the format of content URIs used by ArcVolumeProvider to
content://org.chromium.arc.volumeprovider/<UUID>/*.

This CL adds the support for such URIs to arc.RemovableMedia, while the
support for old content URIs is kept as is. The support for old URIs is
expected to be stopped after we have completely switched the URI format.

BUG=b:164934426
TEST=$ ~/trunk/src/platform/tast/fast_build.sh -T
TEST=Run the following command for DUTs using old/new content URIs:
TEST=$ tast run <DUT> arc.RemovableMedia

Change-Id: Ia9f53f180aa631a94c4795087b8a223c225d5378