tast-tests: add ability to wait for stable location before clicking.
Introduce a way for location to be stable before clicking on items in
the files app. The delay interval for stablity is currently 100ms which
looks to be enough. If necessary, this can be increased for specific
cases where a larger interval is needed. For files app, it can be
increased for a sepcific test or subset of code using SetStablePollOpts.

BUG=b:163101239
TEST=tast run <eve> arc.Drivefs filemanager* crostini.* arc.DownloadsFolder apps.LaunchGallery* policy.DownloadRestrictions ui.HotseatScrollPerf

Change-Id: I94f04d6f0e281d63106a2f14ed1516c29cec2fb8