Fix window resizing scenario of ui.WindowControl
Stainless reveals that the test is failing frequently on certain
devices.

They are failing because the test fails to resize the test
properly, and that happens because of the timing issue of
window bounds right after the mouse release. To stabilize the
result, the only way would be to simply wait for a short
period of time (100msecs) after the mouse is released.

BUG=chromium:1107698
TEST=ui.WindowControl on krane

Change-Id: Ia791aab9f9a54a3f8abc306383d97845a678378b