Fix window resizing scenario of ui.WindowControl
Stainless reveals that the test is failing frequently on certain
devices.

They are failing because the test fails to resize the test
properly. Actually 200msecs drag of shrink/expand is too short,
and sometimes the next click event has the mark of
EF_IS_DOUBLE_CLICK and so the resize session does not start.

This CL simply increases the drag duration to 300msecs. Since
it moves to center and back to the original location, this
will take as 600msecs, which would be long enough to not be
recognized as a double click.

BUG=chromium:1107698
TEST=ui.WindowControl on krane

Change-Id: Ia791aab9f9a54a3f8abc306383d97845a678378b