tast-tests: Improve ephemeral crash collector test flakiness
While the post OOBE, no consent subtest was written as a login test
to allow consent revocation, it does not actually reflect the real
usage of the ephemeral collector, which is run on staritng system
service before the first boot.

Post login, the oobe_completed file is created on a low priority task
which adds flakiness since the file may not exist immediately after
login. Since the ephemeral collector is run before log in, manually
force create the file post-login to emulate the conditions of it's
usage.

BUG=chromium:1122244
TEST=tast run platform.EphemeralCrashCollector.*

Change-Id: I10fbf6a5370b0e5fb00c3a46ab3dde9891582934
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2392147
Tested-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Katherine Threlkeld <kathrelkeld@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Commit-Queue: Sarthak Kukreti <sarthakkukreti@chromium.org>