Add overview animation smoothness for split-view
ui.SplitViewResizePerf is a test to enter/exit split view and
resize the divider, therefore it would be the best to also track
the animation smoothness of overview mode in split-view too.

This CL overhauls the test scenario a bit.  Previously:
1. put a window into left-snapped
2. custom handle the righthand side
3. drag the divider left and right, and moves back to the center
4. repeat the step 3. multiple times

Now:
1. put a window into left-snapped
2. if it is in overview mode at the righthand side, finish the
   overview mode and re-enter into the overview mode again
3. custom handle the righthand side
4. drag the divider left, and moves to the right to exit
   the split screen.
5. repeat the step 1-4 multiple times

That way it can collect the overview animation smoothness in the
same manner.

BUG=chromium:1124081
TEST=ui.SplitViewResizePerf.*

Change-Id: I5ad12cc8e8942f86a3d1c77313878acc6bf37865