Add RV20 Tast-test for WM ARC++ R
Implement RV20: resizable/conversion: portrait.
In RV20 a portrait activity starts in clamshell mode, then tablet
mode enabled. The expectation is to see the display orientation to
change from landscape to portrait. Then tablet mode is disabled, the
display should go back to landscape clamshell. Because of b/162387612
the activity doesn't go back to its original bounds. That should be
addressed after the bug is fixed.
BUG=b:140203139
TEST=tast run $DUT arc.WMResizableConversion

Change-Id: Id0226e14c93bc6aaa2f5598b3ad5649b87821d05