firmware.Config: Allow failure to get model name
Not all devices have a "model", because we support boards that are older
than the idea of a "model".  Most devices have been configured to return
some response for `cros_config / name`, which is how we determine the
model.

But, some devices have not been configured to have a model: for example,
guado. As a result, when our Helper library tries to create a
firmware.Config object for a guado device, it fails to run `cros_config
/ name`, and returns an error.

Our library should recognize that not all devices have a model. If a
model cannot be found, then just set Model="".

BUG=b:161730531
TEST=firmware.Config on guado
TEST=firmware.Config on octopus/bobba360, check USBCInputVoltageLimit=15
TEST=firmware.Config on octopus/apel, check USBCInputVoltageLimit=20

Change-Id: I6e4a6205a44439513fbebe65b022e3b9c88587f2