tast-tests: Replace master, slave in comments, variables and log messages.
As part of cleaning up the code base of offending words, replace all
occurrences of [sS]lave and [mM]aster from comments, local variable
names, and log messages with more appropriate words primary and secondary.

For occurrences that could not be replaced, added a bug and a TODO for
later fixes, (crbug/1126302, crbug/1126921)

See go/coil for details.

Tested using:
 fast_build.sh -T

 tast run localhost:9222
   arc.ContainerMount, arc.StartStop,
   arc.MountOOB, arc.FileSystemPermissions, arc.NativeBridge*

BUG=chromium:1122372
TEST=run Tast tests for all affected files.

Change-Id: I0b0645bacda72122cee9c8b1a9d30e466419e31e