tast-tests: Add hwdep.[No]ForceDischarge to PowerCameraRecordingPerf
On some platforms, force discharging a battery with ectool is not
supported. If so, the battery metric should not be collected.
This CL parameterizes PowerCameraRecordingPerf to decide whether
battery metrics should be collected on a device or not.
It also prevents calling ectool on devices that do not support
force discharge, which leads the test to fail.

BUG=b:157512297
TEST=tast run <DUT w/o battery> arc.PowerCameraRecordingPerf* did not
fail

Change-Id: I54a0eedab0434202c0636bf446413cc022c60683
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2440094
Reviewed-by: Matthias Springer <springerm@chromium.org>
Reviewed-by: Po-Hsien Wang <pwang@chromium.org>
Tested-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Hikaru Nishida <hikalium@chromium.org>