tast-tests: Fix failures of arc.SplitView.clamshell_mode
The tests arc.SplitView.* wait for ARC++ apps to be maximized when they
start, but apparently sometimes, for some reason, the ARC++ apps do not
get maximized when they start in clamshell mode. The present CL simply
removes the code that waits for the ARC++ apps to be maximized when they
start. Only arc.SplitView.clamshell_mode was failing, but the change
applies to all three tests, because the code in question does not seem
to be really necessary for any of them anyway.

BUG=b:166839559
TEST=arc.SplitView.*

Change-Id: I34aa6122fc5a797b1b5378379f8920cc7464bd24