video: PlaybackPerf plays a determined number of frames regardless of video
Originally PlaybackPerf plays a video for a determined time,
approximately 30 seconds. Therefore the number of decoded frames
is dependent on the fps of the test video. It is hard to compare
performance among videos having different fps.
This CL resolves the problem by finishing the video play once the
fixed number of frames is processed, meaning at least the video
progresses, not necessarily decoded. Concretely, PlaybackPerf
plays 30fps video four times and 60fps twice.
This CL shortens the duration for stabilizing cpu load and the
span of measuring gpu and cpu stats to 4 and 12 from 5 and 25
seconds, respectively.
The comparison of measured data is below. They look to not be
affected by this change.

                   | av1 1080p 30fps | av1 1080p 60fps |
                   | before | after  | before | after  |
CPU usage          |  62.44 |  63.62 |  69.91 |  70.42 |
Pkg power usage    |   2.03 |   2.17 |   3.83 |   3.50 |
Dropped frames     |      0 |      0 |     17 |      0 |
Dropped frame rate |      0 |      0 |   0.91 |      0 |

BUG=chromium:1124102
TEST=video.PlaybackPerf.av1_1080p_*fps_sw

Change-Id: I7ef34e49b73602d0e33d4e7d5c9d2d13bba50fd1