wificell: Reconstruct TestFixture when Reinit failed in Pre.
In b/166641939, we find a case that DUT is rebooted in one test for
unknown reason. When this happens, it will break all later WiFi tests
with the same precondition as the gRPC service is broken. Althrough DUT
reboot is unexpected, it is nice to make precondition have some more
tolerance on this.

Also, if we find more failure patterns (e.g. router/pcap disconnect), we
can also add more sanity check in Reinit in the future.

BUG=b:166641939
TEST=tast run $DUT wifi.SimpleConnect.80211*

Change-Id: I79eb692a725d674a2c263237bc4a62890671d294