firmware: Refactor CurrentBootMode as a Reporter
Previously, CurrentBootMode was written as an RPC.
But, reconnecting to RPC after a reboot has been a source of flakiness.
Because we call CurrentBootMode after every mode-aware reboot, this has
meant that all mode-aware reboots are victims of this flakiness.

This CL mitigates that issue by refactoring CurrentBootMode (and
CheckBootMode) to be firmware.Reporter methods instead of RPCs.

BUG=b:155425293
TEST=firmware.BootMode.normal
TEST=firmware.BootMode.rec

Change-Id: I2f8ab0fb7a5c923b512f3f5558fb01430109b365
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2406612
Tested-by: Greg Edelston <gredelston@google.com>
Reviewed-by: Andrew Luo <aluo@chromium.org>
Commit-Queue: Greg Edelston <gredelston@google.com>