tast-tests: Update arc.NativeCrash to check dump files for crash_sender
The previous tests check only for dump files which are directly
created by arc-native-crash-collector{32|64}, which is the first
step of crash reporting. They did so because the later steps
were not implemented yet at the time. This commit updates the
tests to check the final dump files to test all steps of crash
reporting.

Now, some parts of these tests are shared with arc.AppCrash.

BUG=b:136697350
TEST=tast run my_crbook arc.NativeCrash.{mock,real}_consent arc.AppCrash.vm_mock_consent
TEST=tast run my_crbook arc.AppCrash.{mock,real}_consent

Change-Id: Ia31611b3255351de4ae6fb348a98fb4ede810784