arc.VideoEncodeAccel: Add option to download encoded test video.
This CL adds an option to the arc.VideoEncodeAccel.* tests to download
the video encoded by the test, to make debugging more convenient. The
option is disabled by default as we don't want to download the test
video in the lab to limit the size of the test artifacts.

BUG=None
TEST=tast run $HOST arc.VideoEncodeAccel.h264_192p_i420_vm
     on hatch-arc-r

Change-Id: I4e5eeefecfc729180291ee58b6d0922c3e4a1ff3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2408497
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Ricardo Quesada <ricardoq@chromium.org>