Crostini: get installer error messages
Check when an installation fails whether an error message is being
displayed on the screen for the user.  If it is return this string
instead of making developers manually examine screenshots to find the
failure reason.

The main failure reason seen with this is a lack of disk space to
install Crostini.

BUG=None
TEST=Ran on hatch

Change-Id: I50063faa600f4f0976e28a3adee20b3040fdf3c6