Fix error propagation in WaitUntilIdle
This change fixes a shadowing issue that prevented errors from being
returned from WaitUntilIdle. It also changes which context is being
used when polling so that cpu measurement during the final poll
iteration doesn't always timeout.

This change might cause tests which are currently passing to fail.
However, those tests should not have been passing in the first place.
Tests should only be using this function if they actually need the
device to be idle, and this function silently failing was violating
that precondition.

BUG=None
TEST=Run VideoDecodeAccelPerf with locally shorted WaitUntilIdle timeout

Change-Id: I12e00d40122b4061ea9383e3a7daea6258b5dcab