tast-tests: add test for URL popups allow/block lists
This CL adds a tests for PopupsBlockedForUrls and
PopupsAllowedForUrls user policies. The test verifies that
when a URL is specified in the allow list, then it overrides the
default block value from DefaultPopupsSetting policy and popups are
allowed. Similarly when a URL is specified in the block list, then
it overrides the default allow value and popups are blocked.

BUG=chromium:1125587, chromium:1125586
TEST=tast run <ip> policy.PopupsForURLCheck

Change-Id: Icffa74d2566844e5a97092408035551f3b3b8bcf