MemoryChromeOSPerf dump meminfo+zoneinfo
We wanted to know if we are reclaiming in the guest when memory pressure
is high in the host. This change dumps zoneinfo so we can see if there
are any zone free counts below zone low, which would indicate that the
guest is under enough pressure to force continuous reclaim.

Change-Id: Ibdbaef8060b9102fc95c5d35455e319c43dce9a8