wifi: Add another gRPC for waiting for BSSID without scan.
The existing DiscoverBSSID will periodicly trigger scan to discover the
BSS. However, in some tests, we do configure background scan and expect
it to work and detect the BSS. Extract common logic out and provide a
simpler gRPC for the usage.

This will be used in later BgscanBackoff test port.

BUG=b:168681584
TEST=tast run $DUT wifi.RoamDbus

Change-Id: I2de341ee6c0fc1e0c8d67be243beb099813a0b5e