wifi: Add another gRPC for waiting for BSSID without scan.
The existing DiscoverBSSID will periodicly trigger scan to discover the
BSS. However, in some tests, we do configure background scan and expect
it to work and detect the BSS. Extract common logic out and provide a
simpler gRPC for the usage.

This will be used in later BgscanBackoff test port.

BUG=b:168681584
TEST=tast run $DUT wifi.RoamDbus

Change-Id: I2de341ee6c0fc1e0c8d67be243beb099813a0b5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2413828
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Brian Norris <briannorris@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Reviewed-by: Arowa Suliman <arowa@chromium.org>