Ensure vmInstance stops even when the test fails
I've come to notice that crostini.StartupPerf may not stop the
VM instance when the test fails in the middle of the scenario.
This happens consistently on blacktiplte on the failure of
setting up the container.

When this happens, vmInstance.Stop is never called, and therefore
the cros_vm instance remains and consumes 100% of CPU. This can
affect the performance numbers of further tests.

I believe it is better to ensure that the started VM instance
stops at the end of this test.

BUG=chromium:1127525
TEST=crostini.StartupPerf ui.OverviewScrollPerf on blacktiplte

Change-Id: I195bea75154f95a847ad74adcbe4325288fb8947