Sort by metric name on logging the result numbers
Currently it simply iterates over a map, which was good enough,
but now this iteration also logs the result numbers.

Since Go is known to randomize the iteration order of map, this
means that the order in the log for result numbers is different
across the test runs, which is slightly annoying when seeing
the results.  It would be better to sort it by the name of
the metrics.

BUG=none
TEST=ui.WindowControl

Change-Id: I9da6cbd2772c55811627c25dc8595b3e6b1f9b6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2412099
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Tested-by: Jun Mukai <mukai@chromium.org>
Commit-Queue: Jun Mukai <mukai@chromium.org>