tast-tests: Fix ui.StadiaGameplayCUJ issues
WaitForQuiescence after clicking the 'Play' button on Stadia would cause context deadline exceeded 
errors, and it is not necessary because automation API will find the next UI node until the timeout. 
Also it will be safer to check if the window is in full screen before clicking the gameStartButton 
otherwise the clicking might fail.

BUG=chromium:1113012
TEST=tast run IP ui.StadiaGameplayCUJ
Change-Id: I9438b6ffb85bf6bbbb4b0024aad7c84862ea2ce3