wificell: Add Attenuator control
Provides support for controlling Mini-Circuits RC4DAT attenuator.

The support was re-implemented to get rid of telnet in favor of HTTP,
which is simpler to use while proxying on router device (using wget).
Both attenuator host/IP and sshConn handle to proxy device is needed for
using it.

List of fixed attenuations ported from autotest version.

BUG=b:161239563
TEST=AttenuatedPerf(WiP)

Change-Id: I285288907a253db8db790506f7a2a04aa582027a