tast-test: gRPC: Add MonitorShillProperties.
MonitorShillProperties is a bidirectional streaming gRPC. The function
waits for a start signal that has (shill service path) to start the
properties monitor. After, a ready signal is sent to inform the client
that the monitor has been set up successfully. The client should send
a stop signal to stop the monitor and wait to receive a done signal with
the properties changes.

In addition, changed the implementation of the function
AssertNoDisconnection in test_fixture.go to make use of the shill
monitor, instead of using the iw event logger.

BUG=b:166643428
TEST=run the tests:
tast run wifi.SimpleConnect.* [ PASS ]

Change-Id: I41b57c8fabd73adbb8573a576eebd5f9718b0af3