wifi: Port network_WiFi_OverlappingBSSScan to Tast.
Port the test from Autotest to Tast.

The original test parses the logs of hostapd to check the existence of
coexistence action frames, but it is not suggested. Check it with pcap
instead. The frame is expected to be sent multiple times during 100s
period, so we have some tolerance of bad capturing and hopefully won't
make the test too flaky.

BUG=b:169742447
TEST=tast run $DUT wifi.OverlappingBSSScan

Change-Id: Ia2c3427a775c26cdda16b48af14bf000d3bd9277
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2440080
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>