tast-tests: arc: Disable surface scaling hack for betty
There is a hack in betty (the arc_force_2x_scaling USE flag) which
invalidates assumption that 1 Android pixel should match 1 Chrome
devicePixel. This hack only exists for manual testing, so we can
entirely disable this during Tast tests.

The removal of the USE flag was proposed but declined:
https://crrev.com/i/3244437

BUG=b:167341089
TEST=tast run betty arc.DragDrop arc.WindowOutsideDisplay \
arc.WindowDefaultBounds

Change-Id: Ibb0088de90fed7afb9c510d4672a8103a105f1e0