wifi: Don't error when seeing unexpected BSSID.
Although it is not expected in current usage, it might be more general
to allow seeing other BSSID with the same SSID. e.g. a test might
configure two APs and verify the two BSSIDs are all found. It might
fail with the original logic due to the order of BSSAdded.

BUG=b:168681584
TEST=tast run $DUT wifi.RoamDbus

Change-Id: Ifb71123cc730edad8c9672cffce68f0a53e965a5