arc: Add support to cache extracted raw videos in encode tests.
As the raw video files used to test encoding are too large to download
and store on a DUT, videos are stored in the vp9.webm format. Before
each test these videos are extracted to create the raw input video.
This process can take a lot of time, so this CL adds support to cache
the extracted raw video file. This saves about 30s test time when
running 1080p encode testcases. The feature is disabled by default as
we don't want to have DUT's run out of disk space in the lab.

The option is added to both arc++/arcvm variants of these tests:
- arc.VideoEncodeAccel.*
- arc.VideoEncodeAccelPerf.*
- arc.PowerVideoEncodePerf.*

BUG=b:169725788
TEST=various video encode tests on Hatch

Change-Id: Ib1552fd3266b6e2ee33eb0656c590d7a12296161
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2417433
Tested-by: David Staessens <dstaessens@chromium.org>
Commit-Queue: David Staessens <dstaessens@chromium.org>
Reviewed-by: Greg Edelston <gredelston@google.com>