wifi: Enable pcap in BeaconInterval test.
We recently discover that DUTs sometimes spent long time to discover
the service in this test. e.g. [1] takes ~14s while the timeout of
service discovery is 15s. Pcap will help to know if there's some
problems so let's enable it in the test.

[1] https://stainless.corp.google.com/browse/chromeos-autotest-results/swarming-4ebef88395816110/

BUG=None
TEST=tast run $DUT wifi.BeaconInterval

Change-Id: Ife2c07b8ae13ad15f92c07221f810017924966c5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2418277
Tested-by: Yen-lin Lai <yenlinlai@google.com>
Reviewed-by: Shuo-Peng Liao <deanliao@chromium.org>
Commit-Queue: Yen-lin Lai <yenlinlai@google.com>