[Tast] Fix Hotseat animation test
ui.HotseatAnimation is broken because of the feature called
MaintainShelfStateWhenEnteringOverview which was enabled recently. That
feature maintains the hotseat state when entering the overview mode from
the home launcher state. Before the feature was enabled, the hotseat
state became extended. It is why currently the metrics data related to
the extended state are all broken. In addition, the metrics data of
the translucent background animation can only be collected in overflow
mode. However, the current code collects them in non-overflow shelf.

In this CL, disable that feature in the sub-test. Meanwhile, make the
hotseat go to the extended state through gesture swipe up instead of
entering the overview mode. This CL also collects the metrics data
related to the hotseat translucent background only in overflow mode.

BUG=chromium:1129025
TEST=manually

Change-Id: Ic3e88fc289cbe9c2edeb06d3aad973cca1cc5375
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2423188
Tested-by: Andrew Xu <andrewxu@chromium.org>
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: Jun Mukai <mukai@chromium.org>