Make browserWatcher watches the session manager too
Sometimes the session manager may crash during a test scenario,
which will bring a new session manager process, and then later
it will recreate the browser process. There might be some delay
for the relaunch (several seconds), which means the test of
causing the session manager to not fail but the next test case
with the same precondition may suddenly see the chrome crash.

To ensure this situation, this CL makes the browser watcher watch
the session manager's PID.  Once the session manager process
crashes for some reasons, it will set the "browser error" at that
point. That error will be caught when the test scenario attempts
to connect with Chrome, which means the precondition will restart
the UI.

BUG=chromium:1131217
TEST=none

Change-Id: I87c9019bb04e77d8f1eb0e80d218c35f9c9c3c4d