wifi: OptionalDHCPProperties: Don't fail when multiple requests found.
It is possible that the WiFi data frame is re-sent. Or, the UDP packets
is dropped and dhcpcd re-send the packet. In this case, we just want to
make sure that all requests from DUT are with the specific properties.
Relax the count check to fail only when no DHCP request found.

BUG=b:162395242
TEST=tast run $DUT wifi.OptionalDHCPProperties

Change-Id: Ib6ab20665ee18e432ff62d4318643bf3f68e049f