tast-tests: add test for URL images allow/block lists
This CL adds a tests for ImagesBlockedForUrls and
ImagesAllowedForUrls user policies. The test verifies that
when a URL is specified in the allow list, then it overrides the
default block value from DefaultImagesSetting policy and images are
allowed. Similarly when a URL is specified in the block list, then
it overrides the default allow value and images are blocked.

BUG=chromium:1125571, chromium:1125572
TEST=tast run <ip> policy.ImagesForURLCheck.*

Change-Id: Ie852ff7247e077f0f86739c9d04bf0eee49242d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/tast-tests/+/2431517
Reviewed-by: Gabor Zsolt Magda <gabormagda@google.com>
Reviewed-by: Brendan Hansknecht <bhansknecht@chromium.org>
Tested-by: Saurabh Nijhara <snijhara@google.com>
Commit-Queue: Saurabh Nijhara <snijhara@google.com>